% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_transformations.R
\name{comp_transformations}
\alias{comp_transformations}
\alias{comp_to_clr}
\alias{clr_to_comp}
\alias{clr_to_ilr}
\alias{ilr_to_clr}
\alias{comp_to_ilr}
\alias{ilr_to_comp}
\title{Logratio and inverse transformations}
\usage{
comp_to_clr(y)

clr_to_comp(zc)

clr_to_ilr(zc)

ilr_to_clr(zi)

comp_to_ilr(y)

ilr_to_comp(zi)
}
\arguments{
\item{y}{A T x D matrix, where T is the series length and D is 
the simplex dimension. Each row should be a composition.}

\item{zc}{A T x D matrix, where T is the series length and D is 
the simplex dimension. Each row should sum to 0.}

\item{zi}{A T x (D-1) matrix, where T is the series length and D is the simplex dimension.}
}
\value{
Each function returns a matrix with the same number of rows as the input, with row-wise transformed elements.
}
\description{
Collection of functions to return the clr- or ilr-transformed version of a compositional time series, or viceversa. Transformations between ilr and clr forms are included.
}
\details{
Vectors are treated as 1-row matrices, and the returned value is as such.
}
\section{Functions}{
\itemize{
\item \code{comp_to_clr()}: clr transformation of a compositional series.

\item \code{clr_to_comp()}: inverse clr transformation of a series (returns a matrix of compositions).

\item \code{clr_to_ilr()}: ilr transformation of a clr series.

\item \code{ilr_to_clr()}: clr transformation of a ilr series.

\item \code{comp_to_ilr()}: ilr transformation of a compositional series.

\item \code{ilr_to_comp()}: inverse ilr transformation of a series (returns a matrix of compositions).

}}
