% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compStateSpace-package.R
\docType{package}
\name{compStateSpace}
\alias{compStateSpace-package}
\alias{compStateSpace}
\title{compStateSpace: Model-based clustering of compositional time series}
\description{
The package contains functions for the model-based clustering of compositional time series. The method is based on a state-space model representation and on a mixture of experts model for the inclusion of series-specific covariates. The package includes functions for logratio transformation of the series (and their inverse transformations) and functions for Kalman filtering, smoothing and parameter estimation of the elements of the state space representation.
}
