% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mb_clust}
\alias{mb_clust}
\title{Model-based clustering of time series based on a state space representation}
\usage{
mb_clust(
  z_list,
  a0_l,
  P0_l,
  g_T,
  g_Q,
  g_H,
  x_list,
  g_gamma,
  clust_maxit = 1000L,
  clust_toll = 1e-06,
  verbose = 0L
)
}
\arguments{
\item{z_list}{A list of time series. Each element of the list should be a matrix where each row represents a time observation.}

\item{a0_l}{List of initial states (expectations).}

\item{P0_l}{List of variances of the initial states.}

\item{g_T}{List of initial group-specific matrices of state transformation. The matrices should be diagonal.}

\item{g_Q}{List of initial group-specific variance matrices of the state error. The matrices should be diagonal.}

\item{g_H}{List of initial group-specific variance matrices of the observation error. The matrices should be diagonal.}

\item{x_list}{List of series-specific covariates.}

\item{g_gamma}{List of initial values for the component weights parameters.}

\item{clust_maxit}{Maximum number of EM iterations.}

\item{clust_toll}{Tolerance for the relative difference of log-likelihoods criterion to stop the computation.}

\item{verbose}{A non-negative integer to indicate if and how frequently print the likelihood as the EM iterations proceed.}
}
\value{
A list containing:
\describe{
\item{groups_T}{A list with the final group-specific matrices of state transformation.}
\item{groups_Q}{A list with the final group-specific variance matrices of the state error.}
\item{groups_H}{A list with the final group-specific variance matrices of the observation error.}
\item{groups_a0_list}{A list of length K of lists, each of which shows the obtained initial states of each series for the corresponding component.}
\item{groups_P0_list}{A list of length K of lists, each of which shows the variance of the obtained initial states of each series for the corresponding component.}
\item{code}{A value indicating convergence (0) or not (1) of the algorithm.}
\item{clust_it}{Number of performed EM iterations.}
\item{total_lik}{Final complete log-likelihood.}
\item{clust_lik_diff}{Relative log-likelihood difference at the last iteration.}
\item{groups_probs_list}{A list with the probability of each series to belong to each component.}
\item{gamma}{A list with the final component weight parameters.}
\item{time}{Total computational time.}
}
}
\description{
\code{mb_clust} performes the model-based clustering of time series, relying on a state-space representation and a mixture of experts model for the inclusion of series-specific covariates. The method was initially developed for clustering compositional time series according to their evolution in the simplex, by feeding the function their ilr-transformed version.
}
\details{
For compositional time series, the ilr transformation of the series should be provided. \code{z_list <- lapply(y_list, comp_to_ilr)}, where \code{y_list} is a list of matrices with D (= number of parts) columns each, and rows depending on the length of the series, should be enough.

\code{g_T}, \code{g_Q} and \code{g_H} should be lists of length K, where K is the number of groups. Either these, or the values in \code{g_gamma}`, should all be different to ensure the avoidance of identical groups. We recommend providing different starting values for \code{g_T}, \code{g_Q} and \code{g_H}, rather than \code{g_gamma}. A suggested initialization involves the use of a quick clustering procedure (e.g. using \link[dtwclust]{tsclust} from the \code{dtwclust} package) and then some iterations of \link{shumway_EM_list} applied separately to the obtained clusters.

\code{a0_l} and \code{P0_l} should be one of the following: a list (of length K) of lists (each of length equal to the number of series N), or a list of length N. In the former case, the initial states are assumed to be provided for each series for each component. It is useful, for instance, to continue the computation after having obtained a result from the model, if some more iterations are needed. In the latter case, a starting point is assumed to be given for each series, and it is used for the computation for each component. For example, if the initialization with \link{shumway_EM_list} is employed, the obtained results for each series could be used.

\code{x_list} should be of length N, with each covariate vector being series-specific rather than "time-stamp-specific". Every element of the list should be a (P+1)-dimensional vector with 1 as first element, for the intercept.

\code{g_gamma} should be a list of length K, with each element a (P+1)-dimensional vector. The first element is never updated for identifiability reasons, so it should act as reference and a vector of zeros is suggested for it. If the elements in \code{g_T}, \code{g_Q}, and \code{g_H} have been differentiated with an initialization procedure, and no previous knowledge is involved, we suggest to use vectors of zeros for all the component weights parameters, in a sort of "non-informative prior" fashion.

If \code{verbose} is 0, nothing is printed during the computation. If it is a positive integer, the current likelihood and likelihood difference from the previous iteration is printed every \code{verbose} iterations.
}
