% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ShSt_EM}
\alias{ShSt_EM}
\alias{shumway_EM}
\alias{shumway_EM_list}
\title{Estimation of the state-space model system matrices with the EM procedure
from Shumway and Stoffer (1982)}
\usage{
shumway_EM(z, a0, P0, TT, Q, H, EM_toll = 1e-04, EM_maxit = 100L, verbose = 0L)

shumway_EM_list(
  z_list,
  a0_l,
  P0_l,
  TT,
  Q,
  H,
  probs,
  EM_toll = 1e-04,
  EM_maxit = 100L,
  verbose = 0L
)
}
\arguments{
\item{z}{A matrix representing time series. Each row represents a time
observation.}

\item{a0}{Initial state (expectation).}

\item{P0}{Diagonal of the variance matrix of the initial state.}

\item{TT}{Matrix of state transformation. It should be diagonal.}

\item{Q}{Variance matrix of the state error. It should be diagonal.}

\item{H}{Variance matrix of the observation error. It should be diagonal.}

\item{EM_toll}{Tolerance for the relative difference of log-likelihoods criterion to stop the computation.}

\item{EM_maxit}{Maximum number of EM iterations.}

\item{verbose}{A non-negative integer to indicate if and how frequently print 
the current system matrices and likelihood as the EM iterations proceed.}

\item{z_list}{A list of time series. Each element of the list should be a
matrix where each row represents a time observation.}

\item{a0_l}{List of initial states (expectations). Check the `Details`}

\item{P0_l}{List of variances of the initial states. Check the `Details`}

\item{probs}{Weight vector, containing a value in [0,1] for each of the time series. Check the `Details`.}
}
\value{
\code{shumway_EM} returns a list containing:
\describe{
\item{TT}{Estimated matrix of state transformation.}
\item{Q}{Estimated variance matrix of the state error.}
\item{H}{Estimated variance matrix of the observation error.}
\item{a0}{Estimated initial state.}
\item{P0}{Estimated variance of the obtained initial state.}
\item{code}{A value indicating the exit condition of the algorithm. 0 means
convergence, 1 means that the maximum of iterations has been reached,
2 means a decrease (instead of increase) of the log-likelihood, and 3 means
a \code{nan} likelihood.}
\item{EM_it}{Number of performed EM iterations.}
\item{log_lik}{Final marginal log-likelihood.}
\item{diff}{Relative log-likelihood difference at the last iteration.}
}

\code{shumway_EM_list} returns a list containing:
\describe{
\item{TT}{Estimated matrix of state transformation.}
\item{Q}{Estimated variance matrix of the state error.}
\item{H}{Estimated variance matrix of the observation error.}
\item{a0_list}{Estimated initial state.}
\item{P0_list}{Estimated variance of the obtained initial state.}
\item{code}{A value indicating the exit condition of the algorithm. 0 means convergence, 1 means that the maximum of iterations has been reached, 2 means a decrease (instead of increase) of the log-likelihood, and 3 means a \code{nan} likelihood.}
\item{EM_it}{Number of performed EM iterations.}
\item{na_lik_series}{Index of the series that gives the \code{nan} likelihood, if \code{code} = 3. -1 otherwise}
\item{log_lik}{Final weighted marginal log-likelihood.}
\item{diff}{Relative log-likelihood difference at the last iteration.}
}
}
\description{
\code{shumway_EM} and \code{shumway_EM_list} compute the estimates of the system matrices T, Q, H for time series modeled according to the following state-space model:
\deqn{\boldsymbol{\alpha}_t  = \mathbf{T} \boldsymbol{\alpha}_{t-1} + \boldsymbol{\xi}_t\,, \quad \boldsymbol{\xi}_t \sim \mathcal{N}\left(0,\  \mathbf{Q}\right),}
\deqn{\mathbf{z}_t  = \boldsymbol{\alpha}_t + \boldsymbol{\zeta}_t\,, \qquad \boldsymbol{\zeta}_t\sim \mathcal{N}\left(0,\ \mathbf{H}\right).} 
IMPORTANT: T, Q and H are assumed to be diagonal.
}
\details{
For compositional time series, the ilr transformation of the series should be provided. If \code{y_list} is a list of matrices representing compositional times series, with D (= number of parts) columns each, and rows depending on the length of the series, then \code{z_list <- lapply(y_list, \link{comp_to_ilr})}, should be enough.

\code{a0_l} and \code{P0_l} should be a list of length equal to the number of series N. Each element should contain a vector with the same size of a single time-stamp observation: the starting point for \code{a0_l} and the diagonal of the corresponding variance matrix for \code{P0_l}.

In the mixture case, \code{probs} should contain a value between 0 and 1 for each of the time series, corresponding to the probability of the series to belong to the currently considered component. This happens, for example, when \link{mb_clust} is called. If, instead, there are no mixtures involved and the series are all supposed to behave according to the same system matrices, \code{probs} should be a N-dimensional vector of ones.
}
\section{Functions}{
\itemize{
\item \code{shumway_EM()}: Computes the estimate for a single time series.

\item \code{shumway_EM_list()}: computes the estimates for a list of series assumed to 
share the same system matrices. A weight vector \code{probs} can be provided 
if the series do not contribute in the same way to the total likelihood,
for example in a mixture model.

}}
\references{
Shumway, R. H. and Stoffer, D. S. (1982) An approach to time series smoothing and forecasting using the EM algorithm. Journal of Time Series Analysis 3(4), 253–264.

Shumway, R. H. and Stoffer, D. S. (2000) Time Series Analysis and its Applications. Volume 3. Springer
}
