% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ShSt_KFS}
\alias{ShSt_KFS}
\alias{k_filter}
\alias{k_smoother}
\title{Kalman filtering and smoothing of time series}
\usage{
k_filter(z, a0, P0, TT, Q, H)

k_smoother(filtered, TT)
}
\arguments{
\item{z}{A matrix representing time series. Each row represents a time
observation.}

\item{a0}{Initial state (expectation).}

\item{P0}{Diagonal of the variance matrix of the initial state.}

\item{TT}{Matrix of state transformation.}

\item{Q}{The diagonal of the variance matrix of the state error.}

\item{H}{The diagonal of the variance matrix of the observation error.}

\item{filtered}{A list of results from `k_filter`.}
}
\value{
A list containing:
\describe{
\item{alpha_next}{Matrix of predicted estimates \eqn{\mathbf{a}_t^{t-1}}.}
\item{alpha_current}{Matrix of filtered estimates \eqn{\mathbf{a}_t^t}.}
\item{P_next}{Estimated variances of the predicted estimates.}
\item{P_current}{Estimated variances of the filtered estimates.}
\item{log_lik}{Marginal log-likelihood.}
\item{K_n}{Final Kalman gain - useful for smoothing.}
}

\code{k_smoother} returns a list containing:
\describe{
\item{alpha}{Matrix of smoothed estimates \eqn{\mathbf{a}_t^{n}}.}
\item{V}{Estimated variances of the smoothed estimates.}
\item{P_lag}{Lag-One covariance smoother - useful to estimate the system matrices.}
}
}
\description{
\code{k_filter} and \code{k_smoother} compute the filtered and smoothed estimates of time series modeled according to the following state-space model:
\deqn{\boldsymbol{\alpha}_t  = \mathbf{T} \boldsymbol{\alpha}_{t-1} + \boldsymbol{\xi}_t\;,\quad \boldsymbol{\xi}_t \sim \mathcal{N}\left(0,\  \mathbf{Q}\right),}
\deqn{\mathbf{z}_t  = \boldsymbol{\alpha}_t + \boldsymbol{\zeta}_t\,,\qquad \boldsymbol{\zeta}_t\sim \mathcal{N}\left(0,\ \mathbf{H}\right).} 
IMPORTANT: T, Q and H are assumed to be diagonal.
}
\details{
The predicted estimates and their variances are defined as
\deqn{\mathbf{a}_{t}^{t-1} = \mathbb{E}\left(\boldsymbol{\alpha}_{t} \mid \mathbf{Z}_{1:{t-1}}\right),\quad \mathbf{P}_{t}^{t-1} = \mathbb{V}\text{ar}\left(\boldsymbol{\alpha}_{t} \mid \mathbf{Z}_{1:{t-1}}\right),}
where \eqn{\mathbf{Z}_{1:{t-1}}} denotes the series observed up to time \eqn{t-1}. 
The filtered estimates and their variances are defined as
\deqn{\mathbf{a}_{t}^{t} = \mathbb{E}\left(\boldsymbol{\alpha}_{t} \mid \mathbf{Z}_{1:{t}}\right),\quad \mathbf{P}_{t}^{t} = \mathbb{V}\text{ar}\left(\boldsymbol{\alpha}_{t} \mid \mathbf{Z}_{1:{t}}\right).}

The smoothed estimates and their variances are defined as
\deqn{\mathbf{a}_{t}^{n} = \mathbb{E}\left(\boldsymbol{\alpha}_{t} \mid \mathbf{Z}\right),\quad \mathbf{P}_{t}^{n} = \mathbb{V}\text{ar}\left(\boldsymbol{\alpha}_{t} \mid \mathbf{Z}\right),}
that is, having knowledge of the whole observed series.
}
\section{Functions}{
\itemize{
\item \code{k_filter()}: Computes the filtered estimates for a time series.

\item \code{k_smoother()}: Computes the smoothed estimates for a time series, given
the filtered estimates.

}}
\references{
Shumway, R. H. and Stoffer, D. S. (2000) Time Series Analysis and its Applications. 
Volume 3. Springer
}
